+++ src/main/java/br/com/infnet/minddesk/controllers/CategoriaController.java
@@ -3,6 +3,7 @@ package br.com.infnet.minddesk.controllers;
 import br.com.infnet.minddesk.exception.CategoriaException;
 import br.com.infnet.minddesk.model.Categoria;
 import br.com.infnet.minddesk.services.impl.CategoriaServiceImpl;
+import br.com.infnet.minddesk.services.impl.SolicitacaoServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,6 +21,9 @@ public class CategoriaController {
 
     @Autowired
     private CategoriaServiceImpl categoriaService;
+
+    @Autowired
+    private SolicitacaoServiceImpl solicitacaoService;
 
     @Operation(summary = "Adicionar uma Nova Categoria")
     @PostMapping
@@ -68,8 +72,12 @@ public class CategoriaController {
     public ResponseEntity<String> deletarCategoria(@PathVariable Long id) {
         Optional<Categoria> categoriaOptional = categoriaService.findById(id);
         if (categoriaOptional.isPresent()) {
-            categoriaService.deleteById(id);
-            return ResponseEntity.ok("Categoria deletado com sucesso");
+            if (!solicitacaoService.existsByCategoriaId(id)) {
+                categoriaService.deleteById(id);
+                return ResponseEntity.ok("Categoria deletado com sucesso");
+            }else {
+                return ResponseEntity.ok("Não foi possível excluir a Categoria, porque tem pelo menos uma solicitação com essa categoria");
+            }
         } else {
             return ResponseEntity.notFound().build();
         }
