+++ src/main/java/br/com/infnet/minddesk/controllers/ClienteController.java
@@ -2,7 +2,10 @@ package br.com.infnet.minddesk.controllers;
 
 import br.com.infnet.minddesk.exception.ClienteException;
 import br.com.infnet.minddesk.model.Cliente;
+import br.com.infnet.minddesk.records.ClienteDTO;
+import br.com.infnet.minddesk.services.DTOConverterService;
 import br.com.infnet.minddesk.services.impl.ClienteServiceImpl;
+import br.com.infnet.minddesk.services.impl.SolicitacaoServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,10 +24,17 @@ public class ClienteController {
     @Autowired
     private ClienteServiceImpl clienteService;
 
+    @Autowired
+    private DTOConverterService dtoConverterService;
+
+    @Autowired
+    private SolicitacaoServiceImpl solicitacaoService;
+
     @Operation(summary = "Adicionar um Novo Cliente")
     @PostMapping
-    public ResponseEntity<Cliente> criarCliente(@RequestBody Cliente cliente) {
+    public ResponseEntity<Cliente> criarCliente(@RequestBody ClienteDTO dto) {
         try {
+            Cliente cliente = dtoConverterService.converterParaClienteDTO(dto);
             clienteService.save(cliente);
             return ResponseEntity.status(HttpStatus.CREATED).body(cliente);
         }catch (ClienteException e){
@@ -68,8 +78,12 @@ public class ClienteController {
     public ResponseEntity<String> deletarCliente(@PathVariable Long id) {
         Optional<Cliente> clienteOptional = clienteService.findById(id);
         if (clienteOptional.isPresent()) {
-            clienteService.deleteById(id);
-            return ResponseEntity.ok("Cliente deletado com sucesso");
+            if (!solicitacaoService.existsByClienteId(id)) {
+                clienteService.deleteById(id);
+                return ResponseEntity.ok("Cliente deletado com sucesso");
+            }else{
+                return ResponseEntity.ok("Existe pelo menos uma solicitação com este Cliente.");
+            }
         } else {
             return ResponseEntity.notFound().build();
         }
